+++ HISTORY.rst
@@ -8,8 +8,12 @@ Pending Release
 ---------------
 
 * New release notes here
-* Added three built-in condition sets for checking if it's before or after the
-  current date - ``UTCTodayConditionSet``, ``AppTodayConditionSet``, and
+
+1.2.7 (2016-12-01)
+------------------
+
+* Added three built-in condition sets for checking if today is before or after
+  a date - ``UTCTodayConditionSet``, ``AppTodayConditionSet``, and
   ``ActiveTimezoneTodayConditionSet``.
 
 1.2.6 (2016-08-03)